% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{build_initial_df}
\alias{build_initial_df}
\title{Build initial contig-level data frame}
\usage{
build_initial_df(report, contig_names, cov, gc)
}
\arguments{
\item{report}{Data frame. Contains the assembly info and the rownames as the contig names.}

\item{contig_names}{Character vector. Contains the contig IDs.}

\item{cov}{Data frame. Contains the covetage data.}

\item{gc}{Named numeric vector. Contains GC content values, names should match contig IDs.}
}
\value{
A Data frame with contig-level metrics and placeholder columns for later annotation.
}
\description{
Build initial contig-level data frame
}
\keyword{internal}
