% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{build_taxonomic_partitions}
\alias{build_taxonomic_partitions}
\title{Title}
\usage{
build_taxonomic_partitions(tax)
}
\arguments{
\item{tax}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
