% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{build_taxonomy_dataframe}
\alias{build_taxonomy_dataframe}
\title{Title}
\usage{
build_taxonomy_dataframe(taxonomy_list)
}
\arguments{
\item{taxonomy_list}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
