% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_parse_telomeres.R
\name{build_telomere_dataframe}
\alias{build_telomere_dataframe}
\title{Title}
\usage{
build_telomere_dataframe(contigs, left, right)
}
\arguments{
\item{contigs}{TODO}

\item{left}{TODO}

\item{right}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
