% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_incorporate_bam_bubbles.R
\name{choose_best_taxon}
\alias{choose_best_taxon}
\title{Choose the best taxon for a group of contigs}
\usage{
choose_best_taxon(group_contigs, DATA)
}
\arguments{
\item{group_contigs}{Character vector. Names of contigs in a cluster.}

\item{DATA}{Data frame. Main contig data table containing coverage and taxonomic assignments (\code{tax_cons} column).}
}
\value{
Character. The taxon from \code{tax_cons} that best represents the cluster based on coverage similarity.
}
\description{
Choose the best taxon for a group of contigs
}
\keyword{internal}
