% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{cluster_data}
\alias{cluster_data}
\title{Creating and adding cluster classification based on coverage and GC content.}
\usage{
cluster_data(data)
}
\arguments{
\item{data}{Data frame. Contains a contig-level data frame with columns \code{cov}and \code{gc}.}
}
\value{
A data frame with added \code{clas} column indicating cluster membership.
}
\description{
Creating and adding cluster classification based on coverage and GC content.
}
\keyword{internal}
