% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{compute_gc_content}
\alias{compute_gc_content}
\title{Title}
\usage{
compute_gc_content(a)
}
\arguments{
\item{a}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
