% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_export_binned.R
\name{create_output_dirs}
\alias{create_output_dirs}
\title{Create output directories}
\usage{
create_output_dirs(fasta_folder, bed_folder)
}
\arguments{
\item{fasta_folder}{Character string. Path to the directory where binned FASTA files will be saved.}

\item{bed_folder}{Character string. Path to the directory where binned BED files will be saved.}
}
\value{
NULL. Creates two new directories.
}
\description{
Create output directories
}
\keyword{internal}
