% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_incorporate_bam_bubbles.R
\name{detect_cluster}
\alias{detect_cluster}
\title{Title}
\usage{
detect_cluster(shared_matrix, DATA)
}
\arguments{
\item{shared_matrix}{TODO}

\item{DATA}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
