% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{enrich_with_graph_info}
\alias{enrich_with_graph_info}
\title{Annotate contigs with graph completeness}
\usage{
enrich_with_graph_info(data, report)
}
\arguments{
\item{data}{Data frame. Contains a contig-level data frame from build_initial_df.}

\item{report}{Data frame. Contains the assembly info and the rownames as the contig names.}
}
\value{
A modified version of \code{data} with a \code{complete_graph} column where:
\describe{
\item{0}{Path does not contain a \code{"*"}}
\item{1}{Path contains a \code{"*"}}
\item{2}{Path contains a \code{"*"} and is duplicated}
}
}
\description{
Annotate contigs with graph completeness
}
\keyword{internal}
