% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_binned.R
\name{export_binned}
\alias{export_binned}
\title{Title}
\usage{
export_binned(DATA, paths_df, SAMPLE)
}
\arguments{
\item{DATA}{TODO}

\item{paths_df}{TODO}

\item{SAMPLE}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
