% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{extract_taxonomic_labels}
\alias{extract_taxonomic_labels}
\title{Title}
\usage{
extract_taxonomic_labels(data, cl_list)
}
\arguments{
\item{data}{TODO}

\item{cl_list}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
