% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{fetch_tax_classification}
\alias{fetch_tax_classification}
\title{Fetch taxonomy classification from NCBI database}
\usage{
fetch_tax_classification(taxids, api_key = NULL)
}
\arguments{
\item{taxids}{Integer or character vector of taxonomic IDs.}

\item{api_key}{Character string or NULL. API key needed for NCBI Entrez.}
}
\value{
A named list of taxonomic classification.
}
\description{
Fetch taxonomy classification from NCBI database
}
\keyword{internal}
