% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{fill_missing_taxonomy}
\alias{fill_missing_taxonomy}
\title{Fill missing taxonomy ranks by propagating from higher levels}
\usage{
fill_missing_taxonomy(tax)
}
\arguments{
\item{tax}{Data frame. Taxonomy data frame with columns for ranks: domain, kingdom, phylum, etc.}
}
\value{
Taxonomy data frame with missing ranks filled where possible.
}
\description{
Fill missing taxonomy ranks by propagating from higher levels
}
\keyword{internal}
