% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{filter_valid_taxids}
\alias{filter_valid_taxids}
\title{Title}
\usage{
filter_valid_taxids(tax)
}
\arguments{
\item{tax}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
