% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{load_dna_sequences}
\alias{load_dna_sequences}
\title{Loading assembly.fasta file}
\usage{
load_dna_sequences(paths_df)
}
\arguments{
\item{paths_df}{Data frame created with create_filepaths_df}
}
\value{
A matrix of class \code{"DNAbin"} containing DNA sequences
}
\description{
Loading assembly.fasta file
}
\keyword{internal}
