% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{load_taxonomy_table}
\alias{load_taxonomy_table}
\title{Load taxonomy table from a file}
\usage{
load_taxonomy_table(tax_path)
}
\arguments{
\item{tax_path}{Character string. File path to the taxonomy table}
}
\value{
A data frame with columns \code{prot}, \code{taxid}, \code{ev}, and \code{path}.
Duplicated protein IDs are suffixed with \verb{_dup}
}
\description{
Load taxonomy table from a file
}
\keyword{internal}
