% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{merge_tax_data}
\alias{merge_tax_data}
\title{Title}
\usage{
merge_tax_data(tax, tax_df)
}
\arguments{
\item{tax}{TODO}

\item{tax_df}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
