% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_telomeres.R
\name{parse_telomeres}
\alias{parse_telomeres}
\title{Title}
\usage{
parse_telomeres(DATA, paths_df)
}
\arguments{
\item{DATA}{TODO}

\item{paths_df}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
\dontrun{TODO}
}
