% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{process_taxonomy}
\alias{process_taxonomy}
\title{Process taxonomy data through multiple steps}
\usage{
process_taxonomy(tax_path, api_key = NULL)
}
\arguments{
\item{tax_path}{Character string. Contains the path to the taxonomy table.}

\item{api_key}{Character string or NULL. NCBI API key needed for fetching taxonomy classification.}
}
\value{
A list of taxonomic partitions constructed from processed taxonomy data.
}
\description{
Process taxonomy data through multiple steps
}
\keyword{internal}
