% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_parse_telomeres.R
\name{process_telomere_side}
\alias{process_telomere_side}
\title{Process telomere motif hits for one chromosome side}
\usage{
process_telomere_side(file, side)
}
\arguments{
\item{file}{Data frame. Contains columns motif and chrscores, typically from a parsed telomere file.}

\item{side}{Character string. Indicates whether the hits are from the "left" or "right" telomeric region.}
}
\value{
A tibble with columns: \code{motif}, \code{side}, \code{contig}, \code{score}, and \code{rc} (reverse complement).
}
\description{
Process telomere motif hits for one chromosome side
}
\keyword{internal}
