% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{read_busco_file}
\alias{read_busco_file}
\title{Read raw BUSCO data file}
\usage{
read_busco_file(file)
}
\arguments{
\item{file}{Character string. Path to the BUSCO data file}
}
\value{
A filtered data frame of BUSCO entries excluding those with status "Missing".
}
\description{
Read raw BUSCO data file
}
\keyword{internal}
