% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{read_busco_table}
\alias{read_busco_table}
\title{Read and process BUSCO summary table}
\usage{
read_busco_table(file, bacterial = FALSE)
}
\arguments{
\item{file}{Character string. Path to the BUSCO summary file.}

\item{bacterial}{Logical. Whether the dataset is bacterial, affecting sequence name cleaning. Default is FALSE.}
}
\value{
A contingency table summarizing BUSCO counts by sequence and status.
}
\description{
Read and process BUSCO summary table
}
\keyword{internal}
