% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{summarize_busco_by_squence}
\alias{summarize_busco_by_squence}
\title{Summarize BUSCO counts by sequence and status}
\usage{
summarize_busco_by_squence(df)
}
\arguments{
\item{df}{Data frame. BUSCO data frame with sequence and status columns}
}
\value{
A contingency table (matrix) of counts indexed by sequence and status.
}
\description{
Summarize BUSCO counts by sequence and status
}
\keyword{internal}
