% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{add_busco_data}
\alias{add_busco_data}
\title{Title}
\usage{
add_busco_data(data, buscos)
}
\arguments{
\item{data}{wenns_des_is_dann_nochmal_denken}

\item{buscos}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
