% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_contig_table.R
\name{build_contig_table}
\alias{build_contig_table}
\title{Title}
\usage{
build_contig_table(paths_df, API_KEY = API_KEY)
}
\arguments{
\item{paths_df}{TODO}

\item{API_KEY}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
