% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{build_initial_df}
\alias{build_initial_df}
\title{Title}
\usage{
build_initial_df(report, nombres, cov, gc)
}
\arguments{
\item{report}{TODO}

\item{nombres}{TODO}

\item{cov}{TODO}

\item{gc}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
