% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{build_taxonomic_partitions}
\alias{build_taxonomic_partitions}
\title{Build taxonomic partitions for clustering}
\usage{
build_taxonomic_partitions(tax)
}
\arguments{
\item{tax}{Data frame. Taxonomy data frame with contig and rank columns.}
}
\value{
List of \code{cl_partition} objects, one per taxonomic rank.
}
\description{
Build taxonomic partitions for clustering
}
\keyword{internal}
