% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_parse_telomeres.R
\name{build_telomere_dataframe}
\alias{build_telomere_dataframe}
\title{Building telomere data frame with left and right telomeres}
\usage{
build_telomere_dataframe(contigs, left, right)
}
\arguments{
\item{contigs}{Character vector. List of contig names.}

\item{left}{Data frame. Output from process_telomere_side of the left side.}

\item{right}{Data frame. Output from process_telomere_side of the right side.}
}
\value{
A data frame with telomeric motif and score information for each contig side.
}
\description{
Building telomere data frame with left and right telomeres
}
\keyword{internal}
