% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_incorporate_bam_bubbles.R
\name{choose_best_taxon}
\alias{choose_best_taxon}
\title{Title}
\usage{
choose_best_taxon(group_contigs, DATA)
}
\arguments{
\item{group_contigs}{TODO}

\item{DATA}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
