% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{clean_busco_entries}
\alias{clean_busco_entries}
\title{Title}
\usage{
clean_busco_entries(df, bacterial = FALSE)
}
\arguments{
\item{df}{TODO}

\item{bacterial}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
