% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{compute_consensus_clusters}
\alias{compute_consensus_clusters}
\title{Compute consensus cluster classification from cluster list}
\usage{
compute_consensus_clusters(cl_list)
}
\arguments{
\item{cl_list}{List of cluster partitions.}
}
\value{
Vector of factor of consensus cluster class IDs.
}
\description{
Compute consensus cluster classification from cluster list
}
\keyword{internal}
