% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{compute_gc_content}
\alias{compute_gc_content}
\title{Computing GC content}
\usage{
compute_gc_content(dna_seq)
}
\arguments{
\item{dna_seq}{A matrix of class \code{"DNAbin"} containing DNA sequences}
}
\value{
a single numeric value representing the GC content of that sequence
}
\description{
Computing GC content
}
\keyword{internal}
