% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_incorporate_bam_bubbles.R
\name{create_include_ranges}
\alias{create_include_ranges}
\title{Title}
\usage{
create_include_ranges(DATA, telomeres)
}
\arguments{
\item{DATA}{TODO}

\item{telomeres}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
