% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_export_binned.R
\name{create_output_dirs}
\alias{create_output_dirs}
\title{Title}
\usage{
create_output_dirs(fasta_folder, bed_folder)
}
\arguments{
\item{fasta_folder}{TODO}

\item{bed_folder}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
