% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{enrich_with_graph_info}
\alias{enrich_with_graph_info}
\title{Title}
\usage{
enrich_with_graph_info(data, report)
}
\arguments{
\item{data}{TODO}

\item{report}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
