% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_export_binned.R
\name{export_bed_bins}
\alias{export_bed_bins}
\title{Title}
\usage{
export_bed_bins(categories, bed_folder, sample_name)
}
\arguments{
\item{categories}{TODO}

\item{bed_folder}{TODO}

\item{sample_name}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
