% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_export_binned.R
\name{export_fasta_bins}
\alias{export_fasta_bins}
\title{Title}
\usage{
export_fasta_bins(fasta, categories, fasta_folder, sample_name)
}
\arguments{
\item{fasta}{TODO}

\item{categories}{TODO}

\item{fasta_folder}{TODO}

\item{sample_name}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
