% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{extract_taxonomic_labels}
\alias{extract_taxonomic_labels}
\title{Extract taxonomic labels from cluster partitions}
\usage{
extract_taxonomic_labels(data, cl_list)
}
\arguments{
\item{data}{Data frame. Input data with rownames matching cluster partitions.}

\item{cl_list}{List. A list of cluster partition objects, each representing taxonomic clustering at different ranks.}
}
\value{
A data frame with taxonomic classification labels for each input row, with columns for taxonomic ranks.
}
\description{
Extract taxonomic labels from cluster partitions
}
\keyword{internal}
