% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{fetch_tax_classification}
\alias{fetch_tax_classification}
\title{Title}
\usage{
fetch_tax_classification(taxids, API_KEY = NULL)
}
\arguments{
\item{taxids}{TODO}

\item{API_KEY}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
