% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{filter_valid_taxids}
\alias{filter_valid_taxids}
\title{Filter taxonomy entries with valid taxonomic IDs}
\usage{
filter_valid_taxids(tax)
}
\arguments{
\item{tax}{Data frame. Must contain a \code{taxid} column. Typically from parse_protein_ids.}
}
\value{
Filtered data frame with only rows where \code{taxid} is not zero.
}
\description{
Filter taxonomy entries with valid taxonomic IDs
}
\keyword{internal}
