% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{flag_busco_completeness}
\alias{flag_busco_completeness}
\title{Title}
\usage{
flag_busco_completeness(data)
}
\arguments{
\item{data}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
