% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorporate_bam_bubbles.R
\name{incorporate_bam_bubbles}
\alias{incorporate_bam_bubbles}
\title{Title}
\usage{
incorporate_bam_bubbles(DATA, telomeres, paths_df)
}
\arguments{
\item{DATA}{TODO}

\item{telomeres}{TODO}

\item{paths_df}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
\dontrun{TODO}
}
