% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorporate_bam_bubbles.R
\name{incorporate_bam_bubbles}
\alias{incorporate_bam_bubbles}
\title{Title}
\usage{
incorporate_bam_bubbles(DATA, telomeres, paths_df)
}
\arguments{
\item{DATA}{Data frame. The main data dable containing contig information. Row names must be contig names. Typically from build_contig_table.}

\item{telomeres}{Data frame. Telomere information per contig, typically produced by \code{parse_telomeres()}. Must include columns for left and right motifs, scores, and telomere completeness.}

\item{paths_df}{Data frame. Must contain columns \code{name} and \code{path}. Typically from create_filepaths_df.}
}
\value{
TODO
}
\description{
Title
}
\examples{
\dontrun{TODO}
}
