% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{load_assembly_info}
\alias{load_assembly_info}
\title{Loading assembly_info.txt file}
\usage{
load_assembly_info(paths_df)
}
\arguments{
\item{paths_df}{Data frame created with create_filepaths_df}
}
\value{
A data frame with the assembly info and the rownames as the contig names.
}
\description{
Loading assembly_info.txt file
}
\keyword{internal}
