% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{load_coverage_data}
\alias{load_coverage_data}
\title{Loading mean_coverage.tsv file}
\usage{
load_coverage_data(paths_df)
}
\arguments{
\item{paths_df}{Data frame. Typically created with \code{create_filepaths_df}, containing columns \code{name} and \code{path} for locating input files.}
}
\value{
A data frame with the coverage data
}
\description{
Loading mean_coverage.tsv file
}
\keyword{internal}
