% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_incorporate_bam_bubbles.R
\name{load_supplementary_reads}
\alias{load_supplementary_reads}
\title{Title}
\usage{
load_supplementary_reads(bam_file)
}
\arguments{
\item{bam_file}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
