% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{load_taxonomy_table}
\alias{load_taxonomy_table}
\title{Title}
\usage{
load_taxonomy_table(TAX)
}
\arguments{
\item{TAX}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
