% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{merge_tax_data}
\alias{merge_tax_data}
\title{Merge original taxonomy data with classification data frame}
\usage{
merge_tax_data(tax, tax_df)
}
\arguments{
\item{tax}{Data frame. Original taxonomy data including \code{taxid}.}

\item{tax_df}{Data frame. Classification data frame keyed by taxonomic ID.}
}
\value{
Combined data frame with columns from both inputs.
}
\description{
Merge original taxonomy data with classification data frame
}
\keyword{internal}
