% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{parse_protein_ids}
\alias{parse_protein_ids}
\title{Title}
\usage{
parse_protein_ids(tax)
}
\arguments{
\item{tax}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
