% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{process_taxonomy}
\alias{process_taxonomy}
\title{Title}
\usage{
process_taxonomy(TAX, API_KEY = NULL)
}
\arguments{
\item{TAX}{TODO}

\item{API_KEY}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
