% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_parse_telomeres.R
\name{process_telomere_side}
\alias{process_telomere_side}
\title{Title}
\usage{
process_telomere_side(file, side)
}
\arguments{
\item{file}{TODO}

\item{side}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
