% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{read_busco_file}
\alias{read_busco_file}
\title{Title}
\usage{
read_busco_file(file)
}
\arguments{
\item{file}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
A dataframe with busco data and marking empty parts as Missing
}
\keyword{internal}
