% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{read_busco_table}
\alias{read_busco_table}
\title{Title}
\usage{
read_busco_table(file, bacterial = FALSE)
}
\arguments{
\item{file}{TODO}

\item{bacterial}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
