% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_incorporate_bam_bubbles.R
\name{refine_taxonomic_assignments}
\alias{refine_taxonomic_assignments}
\title{Title}
\usage{
refine_taxonomic_assignments(m1_final, DATA)
}
\arguments{
\item{m1_final}{TODO}

\item{DATA}{TODO}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
