% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{summarize_busco_by_squence}
\alias{summarize_busco_by_squence}
\title{Title}
\usage{
summarize_busco_by_squence(df)
}
\arguments{
\item{df}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
