% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_contig_table.R
\name{summarize_taxonomy}
\alias{summarize_taxonomy}
\title{Title}
\usage{
summarize_taxonomy(data)
}
\arguments{
\item{data}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
\keyword{internal}
